Add contact export (day based) With this we can very roughly use this platform to track contacts (without touching writing SQL).
However there is still a lot to be desired. First of all we can only
export the names, however for University and Vienna's contact tracing we
need to add addresses, telephone numbers etc.
Those datapoints are already stored in another service. So either we
integrate with that in this webapp or we write an extern python script
to add those information.

Also it would be nice to select a person that is infected and let the
service automatically figure out with whom they had contact.