Add dark theme The dark theme is automatically selected based on your OS preferences.
Both themes are now based on the Github Theme (primer):
https://primer.style/primitives/colors

The light theme barely changed, but the dark one is completely new.
I am personally not 100% happy with it. While it is definitly better
than nothing, there are also still some quirks and it is visible that
the app was originally designed with the light theme in mind.

It should be good enough for now but problems are:
* Alerts are way too light and don't change depending on the theme
* The gopher in the logo has a black shirt which fuses with the
background in the dark theme which makes the UI look less clean.